rest api's completed

+++ controllers/feed.js
@@ -4,6 +4,48 @@ const path = require('path');
 const { validationResult } = require('express-validator/check');
 
 const Post = require('../models/post');
+const User = require('../models/user')
+
+function errorHandler(message, statusCode) {
+    const error = new Error(message);
+    error.statusCode = statusCode;
+    throw error;
+}
+
+exports.getUserStatus = (req, res, next) => {
+    User
+    .findById(req.userId)
+    .then(user => {
+        if(!user){
+            return errorHandler('no user found', 404);
+        }
+
+        res.status(200).json({ status: user.status });
+    })
+    .catch(err => {
+        next(err);
+    })
+}
+
+exports.updateStatus = (req, res, next) => {
+    const status = req.body.status;
+    User
+    .findById(req.userId)
+    .then(user => {
+        if(!user){
+            return errorHandler('no user found', 404);
+        }
+
+        user.status = status;
+        return user.save();
+    })
+    .then(result => {
+        res.status(200).json({ message: "status updated", status: result.status });
+    })
+    .catch(err => {
+        next(err);
+    })
+}
 
 exports.getPosts = (req, res, next) => {
     const currentPage = req.query.page || 1;
@@ -47,26 +89,33 @@ exports.createPost = (req, res, next) => {
     const imageUrl = req.file.path;
     const title = req.body.title;
     const content = req.body.content;
+    let creator;
 
     const post = new Post({
         title: title,
         content: content,
         imageUrl: imageUrl,
-        creator: { "name": "Mohd Imran" }
+        creator: req.userId
     })
     
     post
     .save()
+    .then(postDoc => {
+        return User.findById(req.userId);
+    })
+    .then(user => {
+        creator = user;
+        user.posts.push(post);
+        return user.save();
+    })
     .then(result => {
         res.status(201).json({
             message: 'Post Created Successfully',
-            post: result
+            post: post,
+            creator: { _id: creator._id, name: creator.name }
         })
     })
     .catch(err => {
-        if(!err.statusCode) {
-            err.statusCode = 500;
-        }
         next(err);
     })
 };
@@ -128,6 +177,10 @@ exports.updatePost = (req, res, next) => {
             throw error;
         }
 
+        if(post.creator.toString() !== req.userId) {
+            return errorHandler('Not Authorized', 403);
+        }
+
         if(imageUrl !== post.imageUrl)
             clearImage(post.imageUrl);
 
@@ -164,11 +217,22 @@ exports.deletePost = (req, res, next) => {
             throw error;
         }
 
-        //check for user login
+
+        if(post.creator.toString() !== req.userId) {
+            return errorHandler('Not Authorized', 403);
+        }
+        
 
         clearImage(post.imageUrl);
         return Post.findByIdAndRemove(postId);
 
+    })
+    .then(result => {
+        return User.findById(req.userId);
+    })
+    .then(user => {
+        user.posts.pull(postId);
+        return user.save();
     })
     .then(result => {
         res
